smaller framebuffer

+++ SDK/HF/Snake_SW/src/lcd_lib/lcd_lib.c
@@ -141,7 +141,7 @@ inline void LcdLineOut(uint8_t data, uint8_t page, uint8_t x){
 	LcdData(data);
 }
 
-void LcdArrayConv(uint16_t *data){
+void LcdArrayConv(uint8_t *data){
 	uint8_t y, x, bit;
 	uint8_t dataout = 0;
