+++ SDK/HF/Snake_SW/src/lcd_lib/lcd_lib.h
@@ -70,7 +70,7 @@ void LcdGoToXY(uint8_t col, uint8_t page);
 void LcdEnable(void);
 void LcdDisable(void);
 void LcdArrayOut(uint8_t *data);
-void LcdArrayConv(uint16_t *data);
+void LcdArrayConv(uint8_t *data);
 inline void LcdLineOut(uint8_t data, uint8_t page, uint8_t x);
 
 #endif
